renamed js file to be inline with master
