+++ test/test_simpleSerialize.js
@@ -4,7 +4,7 @@ const readIntBytes = require('../src/intBytes').readIntBytes;
 const intByteLength = require('../src/intBytes').intByteLength;
 const ActiveState = require('./utils/activeState').ActiveState;
 const AttestationRecord = require('./utils/activeState').AttestationRecord;
-const serialize = require('../src/simpleSerialize').serialize;
+const serialize = require('../src/index').serialize;
 
 describe('SimpleSerialize', () => {
